support_bot/requirements.txt
python-telegram-bot==21.1.1
pymysql==1.1.0
peewee==3.17.5
cryptography==42.0.7